Safari-609.1.11.1/LayoutTests/platform/ios/media/video-controls-captions-trackmenu-expected.txt

Test that we are able to trigger the list of captions, and select from the list.


EVENT(canplaythrough)

*** Set the user language preference.
RUN(internals.setUserPreferredLanguages(['en']))
clickCCButton failed with exception: "undefined: FAIL
END OF TEST
